Moving state around again to appease tests
