+++ src/components/Task.js
@@ -1,7 +1,6 @@
 import React from "react";
 
 function Task({ text, category, onDelete }) {
-
   return (
     <div className="task" >
       <div className="label">{category}</div>
